# docs/api-contract.openapi.yaml
openapi: 3.1.0
info:
  title: Workflow Software Backend API Contract
  version: 0.1.0
  summary: Contract-driven backend API for the workflow builder UI and worker runtime.
  description: |
    This contract is the backend implementation target for the UI. It is based on
    worker/prisma/schema.prisma, the existing worker routes, shared Zod schemas,
    and the current workflow builder screens.

    API responses must use lower_snake_case enum values. Prisma stores generated
    enum values in upper-case form, so backend route serializers should normalize
    values before returning them to the UI.
servers:
  - url: http://localhost:4000
    description: Local worker API default
...
paths:
...
  /api/workflows:
    get:
      tags: [Workflows]
      operationId: listWorkflows
      summary: List workflow drafts for the builder sidebar.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Workflow draft list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Workflows]
      operationId: createWorkflow
      summary: Create a workflow draft from canonical YAML.
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRequest"
      responses:
        "201":
          description: Created workflow draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraft"
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /api/workflows/{workflowId}:
    get:
      tags: [Workflows]
      operationId: getWorkflow
      summary: Get one workflow draft with publish and run summary data.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "200":
          description: Workflow detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/workflows/{workflowId}/draft:
    patch:
      tags: [Workflows]
      operationId: updateWorkflowDraft
      summary: Replace the editable workflow draft YAML.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDraftWorkflowRequest"
      responses:
        "200":
          description: Updated workflow draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDraft"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/workflows/{workflowId}/versions:
    get:
      tags: [Workflow Versions]
      operationId: listWorkflowVersions
      summary: List published immutable versions for one workflow.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Published version list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersionListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/workflow-versions/{workflowId}/publish:
    post:
      tags: [Workflow Versions]
      operationId: publishWorkflowDraft
      summary: Publish the current draft as an immutable workflow version.
      description: |
        This URL matches the currently registered worker route. A future REST cleanup
        may alias this operation to POST /api/workflows/{workflowId}/versions.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "201":
          description: Published workflow version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/workflow-versions/{workflowVersionId}:
    get:
      tags: [Workflow Versions]
      operationId: getWorkflowVersion
      summary: Get one published workflow snapshot.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowVersionId"
      responses:
        "200":
          description: Published workflow version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/runs:
    get:
      tags: [Runs]
      operationId: listWorkflowRuns
      summary: List workflow runs for the history panel.
      x-implementation-status: implemented
      parameters:
        - name: workflowId
          in: query
          required: false
          schema:
            type: string
          description: Filter by logical workflow id.
        - name: workflowVersionId
          in: query
          required: false
          schema:
            type: string
          description: Filter by published workflow version id.
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/WorkflowRunStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Workflow run history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunListResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/runs/workflow-versions/{workflowVersionId}:
    post:
      tags: [Runs]
      operationId: createWorkflowRun
      summary: Create a run for a published workflow version.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkflowVersionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Created workflow run with initialized step runs and code workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/runs/{runId}:
    get:
      tags: [Runs]
      operationId: getWorkflowRun
      summary: Get run detail for status, timeline, inspectors, decisions, and artifacts.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Workflow run detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/runs/{runId}/events:
    get:
      tags: [Runs]
      operationId: listRunEvents
      summary: Poll lightweight run events for live UI updates.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RunId"
        - name: after
          in: query
          required: false
          schema:
            type: string
          description: Return events after this event id or cursor.
      responses:
        "200":
          description: Ordered run events.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RunEvent"
                  nextCursor:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
    WorkflowVersionId:
      name: workflowVersionId
      in: path
      required: true
      schema:
        type: string
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
  responses:
    Error:
      description: Unexpected API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [validation_error, not_found, internal_error]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    WorkflowRunStatus:
      type: string
      enum: [pending, running, waiting, completed, failed, cancelled]
    StepRunStatus:
      type: string
      enum:
        - pending
        - ready
        - running
        - codex_completed
        - waiting_for_codex_permission
        - waiting_for_codex_question
        - waiting_for_human_review
        - waiting_for_evaluator_feedback
        - accepted
        - rejected
        - failed
        - stale
...
    DecisionSource:
      type: string
      enum: [evaluator, human]
...
    WorkflowYaml:
      type: object
      required: [id, name, version, inputs, artifacts, steps, ui]
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        version:
          type: string
          default: 0.1.0
        description:
          type: string
        inputs:
          type: object
          additionalProperties: true
          default: {}
        artifacts:
          type: object
          additionalProperties: true
          default: {}
        steps:
          type: array
          items:
            $ref: "#/components/schemas/StepDefinition"
        ui:
          type: object
          additionalProperties: true
          default: {}
    StepDefinition:
      type: object
      required:
        - id
        - type
        - input_artifacts
        - output_artifacts
        - context_paths
        - tool_capabilities
        - evaluate
        - prompt
        - acceptance
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
        type:
          type: string
          enum: [agent, code_agent]
        upstream:
          type: string
        downstream:
          type: string
        input_artifacts:
          type: array
          items:
            $ref: "#/components/schemas/StepArtifactInput"
          default: []
        output_artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactDefinition"
          default: []
        context_paths:
          type: array
          items:
            $ref: "#/components/schemas/ContextPath"
          default: []
        tool_capabilities:
          type: array
          items:
            type: string
            minLength: 1
          default: []
        evaluate:
          type: object
          required: [evaluator]
          properties:
            evaluator:
              $ref: "#/components/schemas/StepRunEvaluator"
        prompt:
          type: string
          default: ""
        acceptance:
          type: object
          required: [criteria]
          properties:
            criteria:
              type: array
              items:
                type: string
                minLength: 1
              default: []
...
    CreateWorkflowRequest:
      type: object
      required: [draftYaml]
      properties:
        draftYaml:
          $ref: "#/components/schemas/WorkflowYaml"
    UpdateDraftWorkflowRequest:
      type: object
      required: [draftYaml]
      properties:
        draftYaml:
          $ref: "#/components/schemas/WorkflowYaml"
    WorkflowDraft:
      type: object
      required: [id, name, draftYaml, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        draftYaml:
          $ref: "#/components/schemas/WorkflowYaml"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowSummary:
      allOf:
        - $ref: "#/components/schemas/WorkflowDraft"
        - type: object
          properties:
            latestPublishedVersion:
              $ref: "#/components/schemas/WorkflowVersionSummary"
            lastRun:
              $ref: "#/components/schemas/WorkflowRunSummary"
    WorkflowDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowDraft"
        - type: object
          required: [versions, recentRuns]
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowVersionSummary"
            recentRuns:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowRunSummary"
    WorkflowListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowSummary"
        nextCursor:
          type: string
    WorkflowVersion:
      type: object
      required: [id, workflowId, revision, yamlSnapshot, contentHash, publishedAt]
      properties:
        id:
          type: string
        workflowId:
          type: string
        revision:
          type: integer
          minimum: 1
        yamlSnapshot:
          $ref: "#/components/schemas/WorkflowYaml"
        contentHash:
          type: string
        publishedAt:
          type: string
          format: date-time
        lastRun:
          $ref: "#/components/schemas/WorkflowRunSummary"
    WorkflowVersionSummary:
      type: object
      required: [id, workflowId, revision, workflowName, publishedAt, contentHash]
      properties:
        id:
          type: string
        workflowId:
          type: string
        revision:
          type: integer
          minimum: 1
        workflowName:
          type: string
          description: Parsed from yamlSnapshot.name.
        publishedAt:
          type: string
          format: date-time
        contentHash:
          type: string
        lastRun:
          $ref: "#/components/schemas/WorkflowRunSummary"
    WorkflowVersionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowVersion"
        nextCursor:
          type: string
    CreateRunRequest:
      type: object
      properties:
        inputPayload:
          type: object
          additionalProperties: true
          default: {}
    WorkflowRunSummary:
      type: object
      required:
        - id
        - workflowVersionId
        - workflowId
        - workflowName
        - revision
        - status
        - triggerType
        - inputPayload
        - createdAt
        - stepCount
        - completedStepCount
      properties:
        id:
          type: string
        workflowVersionId:
          type: string
        workflowId:
          type: string
        workflowName:
          type: string
        revision:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/WorkflowRunStatus"
        triggerType:
          type: string
          examples: [run_button]
        inputPayload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        stepCount:
          type: integer
          minimum: 0
        completedStepCount:
          type: integer
          minimum: 0
        waitingStepRunId:
          type: string
    WorkflowRunListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowRunSummary"
        nextCursor:
          type: string
    WorkflowRunDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowRunSummary"
        - type: object
          required:
            - workflow
            - stepRuns
            - artifactVersions
            - decisionEvents
            - runEvents
            - codeWorkspaces
          properties:
            workflow:
              $ref: "#/components/schemas/WorkflowYaml"
            stepRuns:
              type: array
              items:
                $ref: "#/components/schemas/StepRunDetail"
            artifactVersions:
              type: array
              items:
                $ref: "#/components/schemas/ArtifactVersion"
            decisionEvents:
              type: array
              items:
                $ref: "#/components/schemas/DecisionEvent"
            runEvents:
              type: array
              items:
                $ref: "#/components/schemas/RunEvent"
            codeWorkspaces:
              type: array
              items:
                $ref: "#/components/schemas/CodeWorkspace"
    StepRunDetail:
      type: object
      required:
        - id
        - workflowRunId
        - stepId
        - attempt
        - status
        - evaluator
        - codexOptions
        - requiresCodeReview
        - createdAt
        - artifactInputs
        - producedArtifacts
        - decisionEvents
        - toolInvocations
        - contextPathEvents
        - codexInteractions
        - codeChangeRecords
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        stepId:
          type: string
        attempt:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/StepRunStatus"
        evaluator:
          $ref: "#/components/schemas/StepRunEvaluator"
        upstreamStepRunId:
          type: string
        downstreamStepRunId:
          type: string
        codexThreadId:
          type: string
        promptSnapshot:
          type: string
        codexOptions:
          type: object
          additionalProperties: true
          default: {}
        codexFinalResponse:
          type: string
        codexUsage:
          type: object
          additionalProperties: true
        codexError:
          type: object
          additionalProperties: true
        codexCompletedAt:
          type: string
          format: date-time
        codeWorkspaceId:
          type: string
        beforeCommit:
          type: string
        afterCommit:
          type: string
        requiresCodeReview:
          type: boolean
        staleReason:
          type: string
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        artifactInputs:
          type: array
          items:
            $ref: "#/components/schemas/StepRunArtifactInput"
        producedArtifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactVersion"
        decisionEvents:
          type: array
          items:
            $ref: "#/components/schemas/DecisionEvent"
        toolInvocations:
          type: array
          items:
            $ref: "#/components/schemas/ToolInvocation"
        contextPathEvents:
          type: array
          items:
            $ref: "#/components/schemas/ContextPathEvent"
        codexInteractions:
          type: array
          items:
            $ref: "#/components/schemas/CodexInteractionEvent"
        codeChangeRecords:
          type: array
          items:
            $ref: "#/components/schemas/CodeChangeRecord"
...
    ArtifactVersion:
      type: object
      required:
        - id
        - workflowRunId
        - artifactKey
        - version
        - producerStepRunId
        - status
        - contentUri
        - contentHash
        - createdAt
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        artifactKey:
          type: string
        version:
          type: integer
          minimum: 1
        displayVersion:
          type: string
          examples: [v1, v2]
        producerStepRunId:
          type: string
        parentVersionId:
          type: string
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        contentUri:
          type: string
        contentHash:
          type: string
        filename:
          type: string
          description: Derived from the workflow snapshot output_artifacts declaration when available.
        format:
          type: string
          enum: [markdown, plain_text]
          description: Derived from the workflow snapshot output_artifacts declaration when available.
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
...
    DecisionEvent:
      type: object
      required: [id, workflowRunId, stepRunId, source, verdict, createdAt]
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        stepRunId:
          type: string
        source:
          $ref: "#/components/schemas/DecisionSource"
        verdict:
          $ref: "#/components/schemas/DecisionVerdict"
        comment:
          type: string
        targetStepId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RunEvent:
      type: object
      required: [id, workflowRunId, eventType, payload, createdAt]
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        eventType:
          type: string
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    CodeWorkspace:
      type: object
      required: [id, workflowRunId, repoPath, baseRef, worktreePath, baseCommit, createdAt, changes]
      properties:
        id:
          type: string
        workflowRunId:
          type: string
        repoPath:
          type: string
        baseRef:
          type: string
        worktreePath:
          type: string
        baseCommit:
          type: string
        createdAt:
          type: string
          format: date-time
        changes:
          type: array
          items:
            $ref: "#/components/schemas/CodeChangeRecord"
...